Day 85 - Watermarking Images with Desktop UI/requirements.txt
Pillow==12.1.1
tkinterdnd2==0.4.2
PyQt5==5.15.11